moved get_data_path to a more appropriate file, as suggested by Siavash
