+++ sepp/filemgr.py
@@ -30,6 +30,7 @@ File and path management.
 
 import os
 import shutil
+import inspect
 
 _LOG = get_logger(__name__)
 
@@ -159,3 +160,36 @@ def directory_has_files_with_prefix(d, prefix):
         if (re.search('^' + prefix + '.*', f) is not None):
             return True
     return False
+
+
+# copy and paste from https://github.com/biocore/scikit-bio/blob/
+# aa502f9dbc7670cf1f8feb6777f6abf181deadc5/skbio/util/_testing.py#L49
+def get_data_path(fn, subfolder='data'):
+    """Return path to filename ``fn`` in the data folder.
+    During testing it is often necessary to load data files. This
+    function returns the full path to files in the ``data`` subfolder
+    by default.
+    Parameters
+    ----------
+    fn : str
+        File name.
+    subfolder : str, defaults to ``data``
+        Name of the subfolder that contains the data.
+    Returns
+    -------
+    str
+        Inferred absolute path to the test data for the module where
+        ``get_data_path(fn)`` is called.
+    Notes
+    -----
+    The requested path may not point to an existing file, as its
+    existence is not checked.
+    """
+    # getouterframes returns a list of tuples: the second tuple
+    # contains info about the caller, and the second element is its
+    # filename
+    callers_filename = inspect.getouterframes(inspect.currentframe())[1][1]
+    path = os.path.dirname(os.path.abspath(callers_filename))
+    data_path = os.path.join(path, subfolder, fn)
+
+    return data_path
