+++ sepp/math_utils.py
@@ -3,8 +3,6 @@ Created on Nov 14, 2012
 
 @author: smirarab
 '''
-import inspect
-import os
 
 
 def gcd(a, b):
@@ -17,36 +15,3 @@ def gcd(a, b):
 def lcm(a, b):
     """Return lowest common multiple."""
     return a * b // gcd(a, b)
-
-
-# copy and paste from https://github.com/biocore/scikit-bio/blob/
-# aa502f9dbc7670cf1f8feb6777f6abf181deadc5/skbio/util/_testing.py#L49
-def get_data_path(fn, subfolder='data'):
-    """Return path to filename ``fn`` in the data folder.
-    During testing it is often necessary to load data files. This
-    function returns the full path to files in the ``data`` subfolder
-    by default.
-    Parameters
-    ----------
-    fn : str
-        File name.
-    subfolder : str, defaults to ``data``
-        Name of the subfolder that contains the data.
-    Returns
-    -------
-    str
-        Inferred absolute path to the test data for the module where
-        ``get_data_path(fn)`` is called.
-    Notes
-    -----
-    The requested path may not point to an existing file, as its
-    existence is not checked.
-    """
-    # getouterframes returns a list of tuples: the second tuple
-    # contains info about the caller, and the second element is its
-    # filename
-    callers_filename = inspect.getouterframes(inspect.currentframe())[1][1]
-    path = os.path.dirname(os.path.abspath(callers_filename))
-    data_path = os.path.join(path, subfolder, fn)
-
-    return data_path
